fix(customization): allow selective override of component options
The code removed in this commit forced the user to specify all options
of the component, even if they only wanted to override a single specific
setting. This case seems to be handled by the `vim.tbl_deep_extend` already
present in this part of the code.

Now the following customization is possible:

```lua
lualine_x = {
  {
    "copilot",
    symbols = {
      -- no need to specify icons, etc.
      status = {
        hl = {
          enabled = "#74c7ec",
        },
      },
    },
  },
  "progress",
  "location",
},
```